Refactored Node to use a `NodeId` trait.
I'm implementing this as a prototype for a design space I don't understand very well. The point of this change is largely to make testing possible. By implementing this simple trait we can use simple identifiers (anything with `Ord + Eq + Copy`) in testing, while using a type containing a `SocketAddr` and consistently-hashed `Key` in the RPC code.